WIP commit - where I left off:
It appears that apollo codegen only generates TS types for queries, mutations, and subscriptions, but am running into challenge creating a simple example of it working.

This may potentially be due to https://github.com/apollographql/apollo-tooling/issues/344#issuecomment-355481287 . I had to downgrade the version of apollo as the newer version was giving a very confusing error message.

Where to go next - try and find a sample project that has this functionality working, as I want to confirm that recent versions of apollo codegen still don't generate types for data types (e.g: the Book in this project)

Resources:
https://youtu.be/ZYhKsX9Alic?t=107